+++ custom_components/spotifyplus/config_flow.py
@@ -53,6 +53,8 @@ from .const import (
     CONF_OPTION_SPOTIFY_SCAN_INTERVAL,
     CONF_OPTION_SPOTIFY_WEBPLAYER_COOKIE_SP_DC,
     CONF_OPTION_SPOTIFY_WEBPLAYER_COOKIE_SP_KEY,
+    CONF_OPTION_TURN_OFF_AUTO_PAUSE,
+    CONF_OPTION_TURN_ON_AUTO_RESUME,
     DEFAULT_OPTION_SPOTIFY_SCAN_INTERVAL,
     DOMAIN, 
     DOMAIN_SCRIPT,
@@ -409,6 +411,8 @@ class SpotifyPlusOptionsFlow(OptionsFlow):
                 self._Options[CONF_OPTION_SCRIPT_TURN_OFF] = user_input.get(CONF_OPTION_SCRIPT_TURN_OFF, None)
                 self._Options[CONF_OPTION_SCRIPT_TURN_ON] = user_input.get(CONF_OPTION_SCRIPT_TURN_ON, None)
                 self._Options[CONF_OPTION_SOURCE_LIST_HIDE] = user_input.get(CONF_OPTION_SOURCE_LIST_HIDE, None)
+                self._Options[CONF_OPTION_TURN_OFF_AUTO_PAUSE] = user_input.get(CONF_OPTION_TURN_OFF_AUTO_PAUSE, True)
+                self._Options[CONF_OPTION_TURN_ON_AUTO_RESUME] = user_input.get(CONF_OPTION_TURN_ON_AUTO_RESUME, True)
                 
                 # validations.
                 # spotify scan interval must be in the 4 to 60 range (if specified).
@@ -467,6 +471,14 @@ class SpotifyPlusOptionsFlow(OptionsFlow):
                                  ): cv.positive_int,
                     vol.Optional(CONF_OPTION_ALWAYS_ON, 
                                  description={"suggested_value": self._Options.get(CONF_OPTION_ALWAYS_ON)},
+                                 ): cv.boolean,
+                    vol.Optional(CONF_OPTION_TURN_OFF_AUTO_PAUSE, 
+                                 description={"suggested_value": self._Options.get(CONF_OPTION_TURN_OFF_AUTO_PAUSE)},
+                                 default=True,  # default to True if not supplied
+                                 ): cv.boolean,
+                    vol.Optional(CONF_OPTION_TURN_ON_AUTO_RESUME, 
+                                 description={"suggested_value": self._Options.get(CONF_OPTION_TURN_ON_AUTO_RESUME)},
+                                 default=True,  # default to True if not supplied
                                  ): cv.boolean,
                 }
             )